dbusif: Don't react to redundant routing messages.
Sometimes (e.g. when muting the mic during calls), policy sends us redundant
routing messages that contain the already active ul and dl routes. There's no
need to do a routing change in this case. It only causes gaps in the audio
because the sink inputs and source outputs are needlessly detached and
re-attached.

(Fixes: NB#279052)