sink-input-ext: Don't classify sink inputs twice.
We're already classifying sink inputs when handling
PA_CORE_HOOK_SINK_INPUT_NEW, so why do it again when handling
PA_CORE_HOOK_SINK_INPUT_PUT? Let's just write the policy group to the sink
input's proplist already in the ..._NEW handler, so we can retrieve it in the
..._PUT handler (it would end up in the proplist anyway).

Also, because it's nowadays possible to specify proplist overwriting for policy
groups (which is done in the ..._NEW handler in order to cut in before
module-stream-restore & co. get their hands on our stream), classifications
relying on proplist values might work incorrectly in the .._PUT handler. A
prime example is the classification of streams into the 'game' group based on
the 'media.role' property. If the 'game' group is specified to overwrite the
'media.role' to 'x-maemo', then the stream will be correctly classified into
'game' in the ..._NEW handler, but then incorrectly reassigned to 'othermedia'
in the ..._PUT handler, because the media role is no longer 'game'. Avoiding
re-classification avoids this problem.