sink-input-ext: Store stream-specific flags into the proplist.
If we need to avoid multiple classifications, we need to store the flags in
addition to the group name.  Previously, it was falsely assumed that the flags
are specific to the group, when in fact they're stream-specific flags.

Recycling stuff through the proplist like this is kind of an ugly hack, but
cleaner ways would require a lot of refactoring.

Fixes: NB#287485