[sink-input-ext] Extend stream corking.
Previously, corking was applied when group with corking enabled was
selected. After this, if the group changed to a group which didn't have
corking enabled, corking was never lifted.

To properly implement both corking and uncorking of streams,
policy-enforcement-module needs to follow the corking state of the
streams it manages so that streams corked by users request aren't
uncorked when the group changes.

As small optimization, if pa_sink_input_ext_cork() is never used, don't
follow streams corking state.